# docs/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "EdgeJobResponse": {
        "additionalProperties": false,
        "description": "Job adquirido pelo Edge Agent.",
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "attempt": {
            "title": "Attempt",
            "type": "integer"
          },
          "competencia": {
            "title": "Competencia",
            "type": "string"
          },
          "fencing_token": {
            "title": "Fencing Token",
            "type": "integer"
          },
          "file_subtype": {
            "title": "File Subtype",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "lease_until": {
            "format": "date-time",
            "title": "Lease Until",
            "type": "string"
          },
          "raw_upload_path": {
            "title": "Raw Upload Path",
            "type": "string"
          },
          "requested_snapshot_mode": {
            "title": "Requested Snapshot Mode",
            "type": "string"
          },
          "source_type": {
            "title": "Source Type",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "agent_id",
          "source_type",
          "file_subtype",
          "competencia",
          "requested_snapshot_mode",
          "fencing_token",
          "attempt",
          "lease_until",
          "raw_upload_path"
        ],
        "title": "EdgeJobResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HeartbeatRequest": {
        "additionalProperties": false,
        "description": "Fence apresentado para renovação do lease.",
        "properties": {
          "fencing_token": {
            "minimum": 0.0,
            "title": "Fencing Token",
            "type": "integer"
          }
        },
        "required": [
          "fencing_token"
        ],
        "title": "HeartbeatRequest",
        "type": "object"
      },
      "HeartbeatResponse": {
        "additionalProperties": false,
        "description": "Lease renovado do job.",
        "properties": {
          "fencing_token": {
            "title": "Fencing Token",
            "type": "integer"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "lease_until": {
            "format": "date-time",
            "title": "Lease Until",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "fencing_token",
          "lease_until"
        ],
        "title": "HeartbeatResponse",
        "type": "object"
      },
...
      "RawUploadResponse": {
        "additionalProperties": false,
        "description": "Recibo do objeto raw persistido.",
        "properties": {
          "object_key": {
            "title": "Object Key",
            "type": "string"
          },
          "object_sha256": {
            "title": "Object Sha256",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "object_key",
          "object_sha256",
          "size_bytes"
        ],
        "title": "RawUploadResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "CnesData Processing Engine",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/edge/jobs/next": {
      "get": {
        "description": "Adquire por CAS o primeiro entre até dez jobs elegíveis.",
        "operationId": "next_job_api_v1_edge_jobs_next_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EdgeJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "204": {
            "description": "No Content"
          }
        },
        "summary": "Next Job",
        "tags": [
          "edge-raw"
        ]
      }
    },
    "/api/v1/edge/jobs/{job_id}/heartbeat": {
      "post": {
        "description": "Renova por 300 segundos o owner e fence apresentados.",
        "operationId": "heartbeat_api_v1_edge_jobs__job_id__heartbeat_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeartbeatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Heartbeat",
        "tags": [
          "edge-raw"
        ]
      }
    },
    "/api/v1/edge/jobs/{job_id}/raw-object": {
      "put": {
        "description": "Transmite um objeto raw sem carregar o payload inteiro em memória.",
        "operationId": "upload_raw_object_api_v1_edge_jobs__job_id__raw_object_put",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Fencing-Token",
            "required": true,
            "schema": {
              "title": "X-Fencing-Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Object-Key",
            "required": true,
            "schema": {
              "title": "X-Object-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Content-Type",
            "required": true,
            "schema": {
              "title": "Content-Type",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RawUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Raw Object",
        "tags": [
          "edge-raw"
        ]
      }
    },
...
  }
}